+++ examples/wallet/travelrule/GetRegionListExample.php
@@ -1,0 +1,18 @@
+<?php
+
+require_once __DIR__.'/../vendor/autoload.php';
+
+use Binance\Client\Wallet\Api\WalletRestApi;
+use Binance\Client\Wallet\WalletRestApiUtil;
+
+function getRegionListExample()
+{
+    $configurationBuilder = WalletRestApiUtil::getConfigurationBuilder();
+    $configurationBuilder->apiKey('apiKey')->privateKey('file:///path/to/private.key');
+    $api = new WalletRestApi($configurationBuilder->build());
+    $countryCode = '';
+    $response = $api->getRegionList($countryCode);
+    print_r($response);
+}
+
+getRegionListExample();